+++ src/Output/Configurator/ConfiguratorInterface.php
@@ -4,11 +4,9 @@ declare(strict_types=1);
 
 namespace Tomaj\NetteApi\Output\Configurator;
 
-use Nette\Application\Request;
-
 interface ConfiguratorInterface
 {
-    public function validateSchema(?Request $request = null): bool;
+    public function validateSchema(): bool;
 
-    public function showErrorDetail(?Request $request = null): bool;
+    public function showErrorDetail(): bool;
 }