+++ src/Output/Configurator/DebuggerConfigurator.php
@@ -4,17 +4,16 @@ declare(strict_types=1);
 
 namespace Tomaj\NetteApi\Output\Configurator;
 
-use Nette\Application\Request;
 use Tracy\Debugger;
 
 class DebuggerConfigurator implements ConfiguratorInterface
 {
-    public function validateSchema(?Request $request = null): bool
+    public function validateSchema(): bool
     {
         return !Debugger::$productionMode;
     }
 
-    public function showErrorDetail(?Request $request = null): bool
+    public function showErrorDetail(): bool
     {
         return !Debugger::$productionMode;
     }